relocate APISearchResponse() from app.records.models to app.search.models
