+++ app/records/models.py
@@ -46,27 +46,6 @@ class APIResponse(APIModel):
         raise Exception("Record template not found in response")
 
 
-class APISearchResponse(APIResponse):
-    @cached_property
-    def records(self) -> list[Record]:
-        records = []
-        if "data" in self._raw:
-            records = [
-                Record(record["@template"]["details"])
-                for record in self._raw["data"]
-                if "@template" in record and "details" in record["@template"]
-            ]
-        return records
-
-    @cached_property
-    def stats_total(self) -> int:
-        return int(self.get("stats.total", "0"))
-
-    @cached_property
-    def stats_results(self) -> int:
-        return int(self.get("stats.results", "0"))
-
-
 class Record(APIModel):
     def __init__(self, raw_data: dict[str, Any]):
         self._raw = raw_data
